Refactor move ItemInterface
