+++ src/Anneck/Game/Item/AbstractItem.php
@@ -14,6 +14,7 @@ namespace Anneck\Game\Item;
 use Anneck\Game\ActionInterface;
 use Anneck\Game\GameInterface;
 use Anneck\Game\GameLogger;
+use Anneck\Game\ItemInterface;
 
 /**
  * The AbstractItem.
