+++ src/Anneck/Game/ItemInterface.php
@@ -9,7 +9,7 @@
  * ************************************************************************
  */
 
-namespace Anneck\Game\Item;
+namespace Anneck\Game;
 
 use Anneck\Game\ActionInterface;
 use Doctrine\Common\Collections\Collection;
